Signup Prob

+++ index.js
@@ -33,7 +33,7 @@ const taskSchema = new mongoose.Schema({
 const Task = mongoose.model("Task", taskSchema);
 app.post("/register", async (req, res) => {
   const { username, password } = req.body;
-  const hashed = await bcrypt(password, 10);
+  const hashed = await bcrypt.hash(password, 10);
   const user = new User({ username, password: hashed });
   await user.save();
   res.json({ message: "Registered" });
@@ -41,9 +41,14 @@ app.post("/register", async (req, res) => {
 app.post("/login", async (req, res) => {
   const { username, password } = req.body;
   const user = await User.findOne({ username });
-  if (!user || (await bcrypt.compare(password, user.password))) {
+  if (!user) {
     return res.status(401).json({ message: "Invalid Credentials" });
   }
+  const valid = await bcrypt.compare(password, user.password);
+  if (!valid) {
+    return res.status(401).json({ message: "Invalid Credentials" });
+  }
+
   const token = jwt.sign({ userId: user._id }, "secret", { expiresIn: "1h" });
   res.json({ token });
 });
@@ -71,7 +76,7 @@ app.post("/tasks", authMiddleware, async (req, res) => {
 });
 //delete task request
 app.delete("/tasks/:id", authMiddleware, async (req, res) => {
-  await Task.findOneAndDelete({ _id: req.params.id.userId }); // ← also check this line
+  await Task.findOneAndDelete({ _id: req.params.id, userId: req.userId });
   res.json({ message: "Task deleted" });
 });
